requirements.txt
# These requirements are only necessary when developing on Oscar.

# development
Werkzeug==0.14.1
django-debug-toolbar==1.10.1
django-extensions==2.1.4
psycopg2>=2.7,<2.8 --no-binary psycopg2

# Sandbox
Pillow==5.3.0
Whoosh==2.7.4
django-environ==0.4.5
django-redis==4.10.0
pysolr==3.7.0
redis==2.10.6
requests==2.20.1
uWSGI==2.0.17.1
whitenoise==4.1.2

# Linting
flake8==3.6.0
flake8-debugger==3.1.0
flake8-blind-except==0.1.1
isort==4.3.4

# Helpers
pyprof2calltree==1.4.4
ipdb==0.11
ipython==7.2.0

# Country data
pycountry==18.5.26

# Wagtail
wagtail>=2.4
wagtail-autocomplete==0.1.1
#wagtailmodelchooser==2.0.0  # now under oscar.vendor.modelchooser
